Exception handling - delete

+++ src/main/java/com/educandoweb/course/services/UserService.java
@@ -4,11 +4,13 @@ import java.util.List;
 import java.util.Optional;
 
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.dao.DataIntegrityViolationException;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
 import com.educandoweb.course.entities.User;
 import com.educandoweb.course.repositories.UserRepository;
+import com.educandoweb.course.services.exceptions.DataBaseException;
 import com.educandoweb.course.services.exceptions.ResourceNotFoundException;
 
 @Service
@@ -33,7 +35,15 @@ public class UserService {
     }
 
     public void delete(Long id){
-        repository.deleteById(id);
+        Optional<User> obj = repository.findById(id);
+        try{
+            if(obj.isPresent())
+                repository.deleteById(id);
+                else
+                    throw new ResourceNotFoundException(id);
+        }catch (DataIntegrityViolationException e){
+            throw new DataBaseException(e.getMessage());
+        }
     }
 
     public User update(Long id, User newData){
